Fixed #3351 - Restructure jetty-unixsocket module.
Split the original module into modules -common, -client, -server.
Added module-info.java for the new modules, referencing the JNR
JPMS module names.
Updated depedencies from old module to new modules.
