+++ jetty-bom/pom.xml
@@ -296,7 +296,17 @@
       </dependency>
       <dependency>
         <groupId>org.eclipse.jetty</groupId>
-        <artifactId>jetty-unixsocket</artifactId>
+        <artifactId>jetty-unixsocket-common</artifactId>
+        <version>10.0.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.jetty</groupId>
+        <artifactId>jetty-unixsocket-client</artifactId>
+        <version>10.0.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.jetty</groupId>
+        <artifactId>jetty-unixsocket-server</artifactId>
         <version>10.0.0-SNAPSHOT</version>
       </dependency>
       <dependency>
