+++ jetty-unixsocket/jetty-unixsocket-common/src/main/java/module-info.java
@@ -1,0 +1,26 @@
+//
+//  ========================================================================
+//  Copyright (c) 1995-2019 Mort Bay Consulting Pty. Ltd.
+//  ------------------------------------------------------------------------
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the Eclipse Public License v1.0
+//  and Apache License v2.0 which accompanies this distribution.
+//
+//      The Eclipse Public License is available at
+//      http://www.eclipse.org/legal/epl-v10.html
+//
+//      The Apache License v2.0 is available at
+//      http://www.opensource.org/licenses/apache2.0.php
+//
+//  You may elect to redistribute this code under either of these licenses.
+//  ========================================================================
+//
+
+module org.eclipse.jetty.unixsocket.common
+{
+    exports org.eclipse.jetty.unixsocket.common;
+
+    requires jnr.unixsocket;
+    requires org.eclipse.jetty.io;
+    requires org.eclipse.jetty.util;
+}
