+++ jetty-unixsocket/jetty-unixsocket-server/src/main/config/etc/jetty-unixsocket-secure.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE Configure PUBLIC "-//Jetty//Configure//EN" "http://www.eclipse.org/jetty/configure_9_3.dtd">
+
+<Configure id="unixSocketHttpConfig" class="org.eclipse.jetty.server.HttpConfiguration">
+  <Call name="addCustomizer">
+    <Arg>
+      <New class="org.eclipse.jetty.server.SecureRequestCustomizer">
+      </New>
+    </Arg>
+  </Call>
+</Configure>
+
