+++ jetty-unixsocket/jetty-unixsocket-server/src/test/resources/jetty-logging.properties
@@ -1,0 +1,3 @@
+org.eclipse.jetty.util.log.class=org.eclipse.jetty.util.log.StdErrLog
+#org.eclipse.jetty.LEVEL=DEBUG
+#org.eclipse.jetty.unixsocket.LEVEL=DEBUG
