--- jetty-unixsocket/src/main/config/etc/jetty-unixsocket.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE Configure PUBLIC "-//Jetty//Configure//EN" "http://www.eclipse.org/jetty/configure_9_3.dtd">
-
-<Configure id="Server" class="org.eclipse.jetty.server.Server">
-  <New id="unixSocketHttpConfig" class="org.eclipse.jetty.server.HttpConfiguration">
-    <Arg><Ref refid="httpConfig"/></Arg>
-  </New>
-
-  <Call name="addConnector">
-    <Arg>
-      <New id="unixSocketConnector" class="org.eclipse.jetty.unixsocket.UnixSocketConnector">
-        <Arg name="server"><Ref refid="Server" /></Arg>
-        <Arg name="selectors" type="int"><Property name="jetty.unixsocket.selectors" default="-1"/></Arg>
-        <Arg name="factories">
-          <Array type="org.eclipse.jetty.server.ConnectionFactory">
-          </Array>
-        </Arg>
-        <Set name="unixSocket"><Property name="jetty.unixsocket" default="/tmp/jetty.sock" /></Set>
-        <Set name="idleTimeout"><Property name="jetty.unixsocket.idleTimeout" default="30000"/></Set>
-        <Set name="acceptQueueSize"><Property name="jetty.unixsocket.acceptQueueSize" default="0"/></Set>
-      </New>
-    </Arg>
-  </Call>
-</Configure>
-
