--- jetty-unixsocket/src/test/resources/jetty-logging.properties
@@ -1,8 +1,0 @@
-org.eclipse.jetty.util.log.class=org.eclipse.jetty.util.log.StdErrLog
-#org.eclipse.jetty.LEVEL=DEBUG
-#org.eclipse.jetty.client.LEVEL=DEBUG
-#org.eclipse.jetty.proxy.LEVEL=DEBUG
-#org.eclipse.jetty.unixsocket.LEVEL=DEBUG
-#org.eclipse.jetty.io.LEVEL=DEBUG
-#org.eclipse.jetty.server.ProxyConnectionFactory.LEVEL=DEBUG
-#org.eclipse.jetty.unixsocket.LEVEL=DEBUG
